feat: マップ画面でのホバーによる画像変化をさらに作りました. 部屋の画面でもコマンドっぽいuiをさらに作りました。
